The world loading process was improved
Entities are once again loaded by a Lua script. This doesn't work on 64
bit because of issues with JavaLua native DLL.

The textures for terrain and sky are now loaded from the database using
the world asset list. This means there is no longer a pause on the ui
thread whilst file io happens.