Autosave system and Physics callbacks
An autosave system has been added which saves the game on a set
interval. This happens on a separate thread so that there is no
noticeable pause in game. At the moment there is no callback when the
save is complete, this would be useful twofold, one, so that feedback
can be given to the player and two, so that no other save operation can
occur whilst another is taking place.